moved tooltip into file
